% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary}
\alias{binary}
\title{Delirium Dataset}
\format{
## `binary`
A data frame with 28 rows and 4 columns:
\describe{
  \item{Study}{Unique study identifier: author, year}
  \item{Events}{Number of events per arm}
  \item{Total}{Total patients per arm}
  \item{Components}{Components per arm seperated by +}
  ...
}
}
\usage{
binary
}
\description{
A Binary Dataset for CNMA of Incidence of Delirium in long format
}
\details{
Derived from (with permision)
Burton JK, Craig L, Yong SQ, Siddiqi N,
Teale EA, Woodhouse R, Barugh AJ, Shepherd AM,
Brunton A, Freeman SC, Sutton AJ, Quinn TJ.
Non‐pharmacological interventions for preventing delirium in
hospitalised non‐ICU patients.
Cochrane Database of Systematic Reviews 2021,
Issue 11. Art. No.: CD013307.
DOI:\url{https://doi.org/10.1002/14651858.CD013307.pub3}
}
\references{
\url{https://doi.org/10.1002/14651858.CD013307.pub3}
}
\author{
Burton JK et al. \email{jenni.burton@glasgow.ac.uk}
}
\keyword{datasets}
