% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_binary_re.R
\name{fit_binary_re}
\alias{fit_binary_re}
\title{fit_binary_re}
\usage{
fit_binary_re(
  data,
  reference_component,
  outcome = "OR",
  chains = 3,
  warmup = 1000,
  iter = 3000,
  seed = 12345,
  max_treedepth = 10,
  adapt_delta = 0.95,
  stepsize = 0.01
)
}
\arguments{
\item{data}{a dataframe containing the required columns see details.}

\item{reference_component}{a character specifying
the control component}

\item{outcome}{(optional) a character specifying the outcome measure.
The default is \code{"OR"}}

\item{chains}{(optional) number of Markov chains.
The Default is 3.}

\item{warmup}{The number of burn-in iterations per chain.
The Default is 500}

\item{iter}{The Number of iterations per chain (including warmup)
The default is 1500}

\item{seed}{The seed for random number generation.
The Default is 12345}

\item{max_treedepth}{(Advanced) See details}

\item{stepsize}{(Advanced) See details}
}
\value{
A fitted stan model of class MetaCNMABayes
}
\description{
Function for fitting a binary random effects CNMA Model using STAN
}
\examples{
\dontrun{
fit_binary_re(binary, "CONTROL")
}
}
