% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCNMABayes-package.R
\docType{package}
\name{MetaCNMABayes}
\alias{MetaCNMABayes-package}
\alias{MetaCNMABayes}
\title{MetaCNMABayes: A package for conducting CNMA using STAN}
\description{
The MetaCNMABayes package provides functions for fitting
and plotting CNMA models using STAN
}
\section{MetaCNMABayes functions}{

The MetaCNMABayes functions ...
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CRSU-Apps/MetaCNMABayes}
  \item Report bugs at \url{https://github.com/CRSU-Apps/MetaCNMABayes/issues}
}

}
\author{
\strong{Maintainer}: Ryan Field \email{ryan.field@glasgow.ac.uk} (\href{https://orcid.org/0000-0002-4424-9890}{ORCID})

Authors:
\itemize{
  \item Suzanne Freeman \email{suzanne.freeman@leicester.ac.uk} (\href{https://orcid.org/0000-0001-8045-4405}{ORCID})
  \item Ellesha Smith \email{eas24@leicester.ac.uk}
  \item Tom Morris \email{tm428@leicester.ac.uk}
}

}
