% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCNMABayes.R
\docType{package}
\name{mypackage}
\alias{mypackage}
\title{MetaCNMABayes: A package for conducting CNMA using STAN}
\description{
The mypackage package provides three categories of important functions:
foo, bar and baz.
}
\section{Mypackage functions}{

The mypackage functions ...
}

