% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MetaCNMABayes.R
\name{plot.MetaCNMABayes}
\alias{plot.MetaCNMABayes}
\title{Default plot function for MetaCNMABayes}
\usage{
\method{plot}{MetaCNMABayes}(obj, ...)
}
\arguments{
\item{obj}{An object of type MetaCNMABayes}

\item{...}{}
}
\description{
Default plot function for MetaCNMABayes
}
\examples{
\dontrun{
model_fit <- fit_binary_fe(binary, "CONTROL")
plot(model_fit)
}
}
