% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.MetaCNMABayes.R
\name{print.summary.MetaCNMABayes}
\alias{print.summary.MetaCNMABayes}
\title{Default print.summary function for MetaCNMABayes}
\usage{
\method{print}{summary.MetaCNMABayes}(obj, ...)
}
\arguments{
\item{obj}{An object of type MetaCNMABayes}

\item{...}{}
}
\description{
Default print.summary function for MetaCNMABayes
}
\examples{
\dontrun{
model_fit <- fit_binary_fe(binary, "CONTROL")
print.summary(model_fit)
}
}
