% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.MetaCNMABayes.R
\name{summary.MetaCNMABayes}
\alias{summary.MetaCNMABayes}
\title{Default summary function for MetaCNMABayes}
\usage{
\method{summary}{MetaCNMABayes}(obj, ...)
}
\arguments{
\item{obj}{An object of type MetaCNMABayes}

\item{...}{}
}
\description{
Default summary function for MetaCNMABayes
}
\examples{
\dontrun{
model_fit <- fit_binary_fe(binary, "CONTROL")
summary(model_fit)
}
}
