% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{continuous}
\alias{continuous}
\title{Length of Stay Dataset}
\format{
## `continuous`
A data frame with 28 rows and 5 columns:
\describe{
  \item{Study}{Unique study identifier: author, year}
  \item{Components}{Components per arm seperated by +}
  \item{Mean}{Mean length of stay per arm}
  \item{SD}{Standard deviation of the mean length of stay per arm}
  \item{Total}{Total patients per arm}
  ...
}
}
\usage{
continuous
}
\description{
A continuous dataset for CNMA of mean length of stay in long format
}
\details{
Data derived from (with permision)

Nicky J. Welton, D. M. Caldwell, E. Adamopoulos, K. Vedhara,
Mixed Treatment Comparison Meta-Analysis of Complex Interventions:
Psychological Interventions in Coronary Heart Disease,
American Journal of Epidemiology,
Volume 169, Issue 9, 1 May 2009,
Pages 1158--1165, \url{https://doi.org/10.1093/aje/kwp014}
}
\references{
\url{https://doi.org/10.1093/aje/kwp014}
}
\author{
Welton NJ et al. \email{Nicky.Welton@bristol.ac.uk}
}
\keyword{datasets}
