% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCNMABayes-package.R
\docType{package}
\name{MetaCNMABayes}
\alias{MetaCNMABayes}
\title{MetaCNMABayes: A package for conducting CNMA using STAN}
\description{
The MetaCNMABayes package provides functions for fitting
and plotting CNMA models using STAN
}
\section{MetaCNMABayes functions}{

The MetaCNMABayes functions ...
}

